+++ src/main/java/ma/vi/esql/database/MariaDb.java
@@ -6,11 +6,11 @@ package ma.vi.esql.database;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
+import ma.vi.base.config.Configuration;
 import ma.vi.esql.translation.Translatable;
 import org.mariadb.jdbc.MariaDbDataSource;
 
 import java.sql.*;
-import java.util.Map;
 import java.util.Properties;
 
 import static ma.vi.base.lang.Errors.unchecked;
@@ -20,11 +20,11 @@ import static ma.vi.esql.translation.Translatable.Target.MARIADB;
  * @author Vikash Madhow (vikash.madhow@gmail.com)
  */
 public class MariaDb extends AbstractDatabase {
-  public MariaDb(Map<String, Object> config) {
+  public MariaDb(Configuration config) {
     Properties props = new Properties();
     props.setProperty("dataSourceClassName", MariaDbDataSource.class.getName());
-    props.setProperty("dataSource.serverName", valueOf(config.getOrDefault(CONFIG_DB_HOST, "localhost")));
-    if (config.containsKey(CONFIG_DB_PORT)) {
+    props.setProperty("dataSource.serverName", valueOf(config.get(CONFIG_DB_HOST, "localhost")));
+    if (config.has(CONFIG_DB_PORT)) {
       props.setProperty("dataSource.portNumber", valueOf(config.get(CONFIG_DB_PORT)));
     }
 
@@ -63,7 +63,7 @@ public class MariaDb extends AbstractDatabase {
       Connection con = DriverManager.getConnection(
           "jdbc:mariadb://"
               + valueOf(config().get(CONFIG_DB_HOST)) + ':'
-              + (config().containsKey(CONFIG_DB_PORT)
+              + (config().has(CONFIG_DB_PORT)
                     ? ':' + valueOf(config().get(CONFIG_DB_PORT))
                     : "")
               + '/' + db,