+++ src/main/java/ma/vi/esql/database/MySql.java
@@ -7,12 +7,12 @@ package ma.vi.esql.database;
 import com.mysql.cj.jdbc.MysqlDataSource;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
+import ma.vi.base.config.Configuration;
 import ma.vi.esql.semantic.type.Type;
 import ma.vi.esql.semantic.type.Types;
 import ma.vi.esql.translation.Translatable;
 
 import java.sql.*;
-import java.util.Map;
 import java.util.Properties;
 
 import static ma.vi.base.lang.Errors.unchecked;
@@ -22,11 +22,11 @@ import static ma.vi.esql.translation.Translatable.Target.MYSQL;
  * @author Vikash Madhow (vikash.madhow@gmail.com)
  */
 public class MySql extends AbstractDatabase {
-  public MySql(Map<String, Object> config) {
+  public MySql(Configuration config) {
     Properties props = new Properties();
     props.setProperty("dataSourceClassName", MysqlDataSource.class.getName());
-    props.setProperty("dataSource.serverName", valueOf(config.getOrDefault(CONFIG_DB_HOST, "localhost")));
-    if (config.containsKey(CONFIG_DB_PORT)) {
+    props.setProperty("dataSource.serverName", valueOf(config.get(CONFIG_DB_HOST, "localhost")));
+    if (config.has(CONFIG_DB_PORT)) {
       props.setProperty("dataSource.portNumber", valueOf(config.get(CONFIG_DB_PORT)));
     }
     props.setProperty("dataSource.databaseName", valueOf(config.get(CONFIG_DB_NAME)));
@@ -63,7 +63,7 @@ public class MySql extends AbstractDatabase {
       Connection con = DriverManager.getConnection(
           "jdbc:mysql://"
               + valueOf(config().get(CONFIG_DB_HOST)) + ':'
-              + (config().containsKey(CONFIG_DB_PORT)
+              + (config().has(CONFIG_DB_PORT)
                     ? ':' + valueOf(config().get(CONFIG_DB_PORT))
                     : "")
               + '/' + valueOf(config().get(CONFIG_DB_NAME)),