+++ src/test/java/ma/vi/esql/Databases.java
@@ -4,9 +4,8 @@
 
 package ma.vi.esql;
 
+import ma.vi.base.config.Configuration;
 import ma.vi.esql.database.*;
-
-import java.util.Map;
 
 import static ma.vi.esql.database.Database.*;
 
@@ -17,7 +16,7 @@ public class Databases {
   public static HSqlDb HSqlDb() {
     if (hSqlDb == null) {
       String userHome = System.getProperty("user.home");
-      hSqlDb = new HSqlDb(Map.of(
+      hSqlDb = new HSqlDb(Configuration.of(
           CONFIG_DB_NAME, "file:" + userHome + "/testdb/data",
           CONFIG_DB_USER, "SA",
           CONFIG_DB_PASSWORD, "",
@@ -28,7 +27,7 @@ public class Databases {
 
   public static Postgresql Postgresql() {
     if (postgresql == null) {
-      postgresql = new Postgresql(Map.of(
+      postgresql = new Postgresql(Configuration.of(
           CONFIG_DB_NAME, "test",
           CONFIG_DB_USER, "test",
           CONFIG_DB_PASSWORD, "test",
@@ -40,7 +39,7 @@ public class Databases {
 
   public static SqlServer SqlServer() {
     if (sqlServer == null) {
-      sqlServer = new SqlServer(Map.of(
+      sqlServer = new SqlServer(Configuration.of(
           CONFIG_DB_NAME, "test",
           CONFIG_DB_USER, "test",
           CONFIG_DB_PASSWORD, "test",
@@ -52,7 +51,7 @@ public class Databases {
 
   public static MariaDb MariaDb() {
     if (mariaDb == null) {
-      mariaDb = new MariaDb(Map.of(
+      mariaDb = new MariaDb(Configuration.of(
           CONFIG_DB_NAME, "test",
           CONFIG_DB_USER, "test",
           CONFIG_DB_PASSWORD, "test",